requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.8
attrs==25.1.0
black==24.10.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
coverage==7.6.10
cymem==2.0.11
dill==0.3.9
distro==1.9.0
dramatiq==1.17.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.10.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
murmurhash==1.0.12
mypy==1.14.1
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
openai==1.68.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
packaging==24.2
pathable==0.4.4
pathspec==0.12.1
pgvector==0.4.0
pillow==11.1.0
pip-autoremove==0.10.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.21.1
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyHamcrest==2.1.0
pylint==3.3.3
pylint-module-boundaries==1.3.1
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.2
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.1.2
yarl==1.18.3
zipp==3.21.0
